Rename mqtt_operation_converter to tedge_operation_converter
