+++ crates/core/tedge_agent/src/main.rs
@@ -8,11 +8,11 @@ use tedge_config::DEFAULT_TEDGE_CONFIG_PATH;
 mod agent;
 mod error;
 mod file_transfer_server;
-mod mqtt_operation_converter;
 mod restart_manager;
 mod software_list_manager;
 mod software_update_manager;
 mod state_repository;
+mod tedge_operation_converter;
 
 #[derive(Debug, clap::Parser)]
 #[clap(
